+++ tests/VocabularyCoach.Services.UnitTests/Helpers/CheckResultsExtensions.cs
@@ -5,13 +5,13 @@ namespace VocabularyCoach.Services.UnitTests.Helpers
 {
 	internal static class CheckResultsExtensions
 	{
-		public static StudiedText ToStudiedText(this IEnumerable<CheckResult> checkResults, string itemId)
+		public static StudiedText ToStudiedText(this IEnumerable<CheckResult> checkResults)
 		{
 			return new StudiedText(checkResults)
 			{
 				TextInStudiedLanguage = new LanguageText
 				{
-					Id = new ItemId(itemId),
+					Id = new ItemId("test"),
 				},
 			};
 		}