+++ tests/VocabularyCoach.Services.UnitTests/Helpers/DateTimeExtensions.cs
@@ -1,0 +1,26 @@
+using System;
+using VocabularyCoach.Models;
+
+namespace VocabularyCoach.Services.UnitTests.Helpers
+{
+	internal static class DateTimeExtensions
+	{
+		public static CheckResult ToSuccessfulCheckResult(this DateTime dateTime)
+		{
+			return new CheckResult
+			{
+				DateTime = dateTime,
+				CheckResultType = CheckResultType.Ok,
+			};
+		}
+
+		public static CheckResult ToFailedCheckResult(this DateTime dateTime)
+		{
+			return new CheckResult
+			{
+				DateTime = dateTime,
+				CheckResultType = CheckResultType.Misspelled,
+			};
+		}
+	}
+}