+++ tests/VocabularyCoach.Services.UnitTests/Internal/TextsForPracticeSelectorTests.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Globalization;
 using System.Linq;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Moq;
 using Moq.AutoMock;
 using VocabularyCoach.Models;
 using VocabularyCoach.Services.Internal;
@@ -9,11 +11,9 @@ using VocabularyCoach.Services.UnitTests.Helpers;
 
 namespace VocabularyCoach.Services.UnitTests.Internal
 {
-	// TODO: Restore UT for TextsForPracticeSelector.
 	[TestClass]
 	public class TextsForPracticeSelectorTests
 	{
-/*
 		[TestMethod]
 		public void GetTextsForPractice_ForMissingStudiedTexts_ReturnsEmptyCollection()
 		{
@@ -22,11 +22,14 @@ namespace VocabularyCoach.Services.UnitTests.Internal
 			var studiedTexts = Array.Empty<StudiedText>();
 
 			var mocker = new AutoMocker();
+			mocker.GetMock<INextCheckDateProvider>()
+				.Setup(x => x.GetNextCheckDate(It.IsAny<StudiedText>())).Returns(new DateOnly(2023, 07, 11));
+
 			var target = mocker.CreateInstance<TextsForPracticeSelector>();
 
 			// Act
 
-			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts);
+			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts, 100);
 
 			// Assert
 
@@ -34,368 +37,306 @@ namespace VocabularyCoach.Services.UnitTests.Internal
 		}
 
 		[TestMethod]
-		public void GetTextsForPractice_ForTextsWithNoChecks_ReturnsSuchTexts()
+		public void GetTextsForPractice_ForTextsWithNextCheckDateNotYetReached_DoesNotReturnSuchTexts()
 		{
 			// Arrange
 
+			var mocker = new AutoMocker();
+
 			var studiedTexts = new[]
 			{
-				Array.Empty<CheckResult>().ToStudiedText("Text with no checks"),
+				CreateStudiedText("Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 12)),
+				CreateStudiedText("Text 3", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 4", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 12)),
 			};
 
-			var mocker = new AutoMocker();
 			var target = mocker.CreateInstance<TextsForPracticeSelector>();
 
 			// Act
 
-			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts);
-
-			// Assert
-
-			result.Should().BeEquivalentTo(studiedTexts, x => x.WithoutStrictOrdering());
-		}
-
-		[TestMethod]
-		public void GetTextsForPractice_ForTextsWithSmallNumberOfAllSuccessfulChecks_TreatsFirstMissingCheckAsFailed()
-		{
-			// Arrange
-
-			var studiedTexts = new[]
-			{
-				new[]
-				{
-					// Next check: +2 days - 2023.07.21
-					GetSuccessfulCheck(new DateTime(2023, 07, 19)),
-				}.ToStudiedText("Text with 1 successful check - returned"),
-
-				new[]
-				{
-					// Next check: +2 days - 2023.07.22
-					GetSuccessfulCheck(new DateTime(2023, 07, 20)),
-				}.ToStudiedText("Text with 1 successful check - not returned"),
-
-				new[]
-				{
-					// Next check: +3 days - 2023.07.21
-					GetSuccessfulCheck(new DateTime(2023, 07, 17)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 18)),
-				}.ToStudiedText("Text with 2 successful checks - returned"),
-
-				new[]
-				{
-					// Next check: +3 days - 2023.07.22
-					GetSuccessfulCheck(new DateTime(2023, 07, 18)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 19)),
-				}.ToStudiedText("Text with 2 successful checks - not returned"),
-
-				new[]
-				{
-					// Next check: +7 days - 2023.07.21
-					GetSuccessfulCheck(new DateTime(2023, 07, 12)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 13)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 14)),
-				}.ToStudiedText("Text with 3 successful checks - returned"),
-
-				new[]
-				{
-					// Next check: +7 days - 2023.07.22
-					GetSuccessfulCheck(new DateTime(2023, 07, 13)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 14)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 15)),
-				}.ToStudiedText("Text with 3 successful checks - not returned"),
-
-				new[]
-				{
-					// Next check: +14 days - 2023.07.21
-					GetSuccessfulCheck(new DateTime(2023, 07, 04)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 05)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 07)),
-				}.ToStudiedText("Text with 4 successful checks - returned"),
-
-				new[]
-				{
-					// Next check: +14 days - 2023.07.22
-					GetSuccessfulCheck(new DateTime(2023, 07, 05)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 07)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 08)),
-				}.ToStudiedText("Text with 4 successful checks - not returned"),
-			};
-
-			var mocker = new AutoMocker();
-			var target = mocker.CreateInstance<TextsForPracticeSelector>();
-
-			// Act
-
-			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 21), studiedTexts);
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts, 100);
 
 			// Assert
 
 			var expectedTextsForCheck = new[]
 			{
-				studiedTexts[0],
-				studiedTexts[2],
-				studiedTexts[4],
-				studiedTexts[6],
+				studiedTexts.Get("Text 1"),
+				studiedTexts.Get("Text 3"),
 			};
 
-			result.Should().BeEquivalentTo(expectedTextsForCheck, x => x.WithoutStrictOrdering());
+			textsForPractice.Should().BeEquivalentTo(expectedTextsForCheck, x => x.WithoutStrictOrdering());
 		}
 
 		[TestMethod]
-		public void GetTextsForPractice_ForTextsWithAllSuccessfulChecks_AppliesCorrectCheckInterval()
+		public void GetTextsForPractice_IfDailyLimitIsReached_ReturnsEmptyCollection()
 		{
 			// Arrange
 
+			var mocker = new AutoMocker();
+
 			var studiedTexts = new[]
 			{
-				new[]
-				{
-					// Next check: +30 days - 2023.07.11
-					GetSuccessfulCheck(new DateTime(2023, 06, 07)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 08)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 09)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 10)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 11)),
-				}.ToStudiedText("Text with 5 successful checks - returned"),
-
-				new[]
-				{
-					// Next check: +30 days - 2023.07.12
-					GetSuccessfulCheck(new DateTime(2023, 06, 08)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 09)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 10)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 11)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 12)),
-				}.ToStudiedText("Text with 5 successful checks - not returned"),
-
-				new[]
-				{
-					// Next check: +30 days - 2023.07.11
-					GetSuccessfulCheck(new DateTime(2023, 06, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 07)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 08)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 09)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 10)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 11)),
-				}.ToStudiedText("Text with 6 successful checks - returned"),
-
-				new[]
-				{
-					// Next check: +30 days - 2023.07.11
-					GetFailedCheck(new DateTime(2023, 06, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 07)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 08)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 09)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 10)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 11)),
-				}.ToStudiedText("Text with 5 successful checks and old failed check - returned"),
-
-				// This case was added to verify a fix for ArgumentOutOfRangeException when there are 7 or more check results for a text.
-				new[]
-				{
-					// Next check: +30 days - 2023.07.11
-					GetFailedCheck(new DateTime(2023, 06, 05)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 07)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 08)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 09)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 10)),
-					GetSuccessfulCheck(new DateTime(2023, 06, 11)),
-				}.ToStudiedText("Text with 7 successful checks - returned"),
+				CreateStudiedText("Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 3", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
 			};
 
-			var mocker = new AutoMocker();
 			var target = mocker.CreateInstance<TextsForPracticeSelector>();
 
 			// Act
 
-			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts);
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts, 2);
+
+			// Assert
+
+			textsForPractice.Should().BeEmpty();
+		}
+
+		[TestMethod]
+		public void GetTextsForPractice_IfSomeTextsWerePracticedToday_TakesSuchTextsIntoAccountForDailyLimit()
+		{
+			// Arrange
+
+			var mocker = new AutoMocker();
+
+			var studiedTexts = new[]
+			{
+				CreateStudiedText("Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 12)),
+				CreateStudiedText("Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 12)),
+				CreateStudiedText("Text 3", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 4", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 5", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 6", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 7", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+			};
+
+			var target = mocker.CreateInstance<TextsForPracticeSelector>();
+
+			// Act
+
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts, 5);
+
+			// Assert
+
+			textsForPractice.Count.Should().Be(3);
+
+			textsForPractice.Should().NotContain(studiedTexts.Get("Text 1"));
+			textsForPractice.Should().NotContain(studiedTexts.Get("Text 2"));
+		}
+
+		[TestMethod]
+		public void GetTextsForPractice_IfNumberOfSuitablePreviouslyPracticedTextsExceedsDailyLimit_ReturnsTextsWithMostOverduePracticeDate()
+		{
+			// Arrange
+
+			var mocker = new AutoMocker();
+
+			var studiedTexts = new[]
+			{
+				CreateStudiedText("Group 3 - Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 14)),
+				CreateStudiedText("Group 3 - Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 14)),
+				CreateStudiedText("Group 3 - Text 3", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 14)),
+
+				CreateStudiedText("Group 1 - Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 12)),
+				CreateStudiedText("Group 1 - Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 12)),
+				CreateStudiedText("Group 1 - Text 3", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 12)),
+
+				CreateStudiedText("Group 2 - Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 13)),
+				CreateStudiedText("Group 2 - Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 13)),
+				CreateStudiedText("Group 2 - Text 3", mocker, lastCheckDate: new DateTime(2023, 07, 11), nextCheckDate: new DateOnly(2023, 07, 13)),
+			};
+
+			var target = mocker.CreateInstance<TextsForPracticeSelector>();
+
+			// Act
+
+			var textsForPractice1 = target.GetTextsForPractice(new DateOnly(2023, 07, 14), studiedTexts, 3);
+			var textsForPractice2 = target.GetTextsForPractice(new DateOnly(2023, 07, 14), studiedTexts, 6);
+
+			// Assert
+
+			var expectedTextsForCheck1 = new[]
+			{
+				studiedTexts.Get("Group 1 - Text 1"),
+				studiedTexts.Get("Group 1 - Text 2"),
+				studiedTexts.Get("Group 1 - Text 3"),
+			};
+
+			var expectedTextsForCheck2 = new[]
+			{
+				studiedTexts.Get("Group 1 - Text 1"),
+				studiedTexts.Get("Group 1 - Text 2"),
+				studiedTexts.Get("Group 1 - Text 3"),
+				studiedTexts.Get("Group 2 - Text 1"),
+				studiedTexts.Get("Group 2 - Text 2"),
+				studiedTexts.Get("Group 2 - Text 3"),
+			};
+
+			textsForPractice1.Should().BeEquivalentTo(expectedTextsForCheck1, x => x.WithoutStrictOrdering());
+			textsForPractice2.Should().BeEquivalentTo(expectedTextsForCheck2, x => x.WithoutStrictOrdering());
+		}
+
+		[TestMethod]
+		public void GetTextsForPractice_IfNumberOfSuitablePreviouslyPracticedTextsForLatestCheckDateExceedsDailyLimit_ReturnsRandomSubsetOfSuchTexts()
+		{
+			// Arrange
+
+			var mocker = new AutoMocker();
+
+			// Since texts are picked in random order, we take quite a big number of such texts,
+			// so that we can check that items are actually randomized, i.e. texts are returned from both halves.
+			var studiedTexts1 = Enumerable.Range(1, 100)
+				.Select(n => CreateStudiedText($"Practiced Text - 1st Half - {n}", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)))
+				.ToList();
+
+			var studiedTexts2 = Enumerable.Range(1, 100)
+				.Select(n => CreateStudiedText($"Practiced Text - 2nd Half - {n}", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)))
+				.ToList();
+
+			var target = mocker.CreateInstance<TextsForPracticeSelector>();
+
+			// Act
+
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts1.Concat(studiedTexts2), 100);
+
+			// Assert
+
+			textsForPractice.Should().Contain(x => studiedTexts1.Any(y => ReferenceEquals(x, y)));
+			textsForPractice.Should().Contain(x => studiedTexts2.Any(y => ReferenceEquals(x, y)));
+		}
+
+		[TestMethod]
+		public void GetTextsForPractice_IfNumberOfSuitablePreviouslyPracticedTextsDoesNotExceedDailyLimit_ReturnsFirstUnpracticedTexts()
+		{
+			// Arrange
+
+			var mocker = new AutoMocker();
+
+			var studiedTexts = new[]
+			{
+				CreateStudiedText("Text 1", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+				CreateStudiedText("Text 2", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11)),
+
+				CreateStudiedTextWithNoChecks("Text 3", mocker, createDate: new DateTime(2023, 07, 03)),
+				CreateStudiedTextWithNoChecks("Text 4", mocker, createDate: new DateTime(2023, 07, 05)),
+				CreateStudiedTextWithNoChecks("Text 5", mocker, createDate: new DateTime(2023, 07, 02)),
+				CreateStudiedTextWithNoChecks("Text 6", mocker, createDate: new DateTime(2023, 07, 01)),
+				CreateStudiedTextWithNoChecks("Text 7", mocker, createDate: new DateTime(2023, 07, 04)),
+			};
+
+			var target = mocker.CreateInstance<TextsForPracticeSelector>();
+
+			// Act
+
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2023, 08, 11), studiedTexts, 5);
 
 			// Assert
 
 			var expectedTextsForCheck = new[]
 			{
-				studiedTexts[0],
-				studiedTexts[2],
-				studiedTexts[3],
-				studiedTexts[4],
+				studiedTexts.Get("Text 1"),
+				studiedTexts.Get("Text 2"),
+				studiedTexts.Get("Text 3"),
+				studiedTexts.Get("Text 5"),
+				studiedTexts.Get("Text 6"),
 			};
 
-			result.Should().BeEquivalentTo(expectedTextsForCheck, x => x.WithoutStrictOrdering());
+			textsForPractice.Should().BeEquivalentTo(expectedTextsForCheck, x => x.WithoutStrictOrdering());
 		}
 
 		[TestMethod]
-		public void GetTextsForPractice_ForTextsWithFailedCheck_AppliesCorrectCheckInterval()
+		public void GetTextsForPractice_IfOnlyTextsPracticedBeforeAreReturned_ReturnsTextsInRandomOrder()
 		{
 			// Arrange
 
-			var studiedTexts = new[]
-			{
-				new[]
-				{
-					// Next check: +1 day - 2023.07.21
-					GetFailedCheck(new DateTime(2023, 07, 20)),
-				}.ToStudiedText("Text with 1st failed check - returned"),
-
-				new[]
-				{
-					// Next check: +1 day - 2023.07.22
-					GetFailedCheck(new DateTime(2023, 07, 21)),
-				}.ToStudiedText("Text with 1st failed check - not returned"),
-
-				new[]
-				{
-					// Next check: +2 days - 2023.07.21
-					GetFailedCheck(new DateTime(2023, 07, 18)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 19)),
-				}.ToStudiedText("Text with 2nd failed check - returned"),
-
-				new[]
-				{
-					// Next check: +2 days - 2023.07.22
-					GetFailedCheck(new DateTime(2023, 07, 19)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 20)),
-				}.ToStudiedText("Text with 2nd failed check - not returned"),
-
-				new[]
-				{
-					// Next check: +3 days - 2023.07.21
-					GetFailedCheck(new DateTime(2023, 07, 16)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 17)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 18)),
-				}.ToStudiedText("Text with 3rd failed check - returned"),
-
-				new[]
-				{
-					// Next check: +3 days - 2023.07.22
-					GetFailedCheck(new DateTime(2023, 07, 17)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 18)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 19)),
-				}.ToStudiedText("Text with 3rd failed check - not returned"),
-
-				new[]
-				{
-					// Next check: +7 days - 2023.07.21
-					GetFailedCheck(new DateTime(2023, 07, 11)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 12)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 13)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 14)),
-				}.ToStudiedText("Text with 4th failed check - returned"),
-
-				new[]
-				{
-					// Next check: +7 days - 2023.07.22
-					GetFailedCheck(new DateTime(2023, 07, 12)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 13)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 14)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 15)),
-				}.ToStudiedText("Text with 4th failed check - not returned"),
-
-				new[]
-				{
-					// Next check: +14 days - 2023.07.21
-					GetFailedCheck(new DateTime(2023, 07, 03)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 04)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 05)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 07)),
-				}.ToStudiedText("Text with 5th failed check - returned"),
-
-				new[]
-				{
-					// Next check: +14 days - 2023.07.22
-					GetFailedCheck(new DateTime(2023, 07, 04)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 05)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 06)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 07)),
-					GetSuccessfulCheck(new DateTime(2023, 07, 08)),
-				}.ToStudiedText("Text with 5th failed check - not returned"),
-			};
-
 			var mocker = new AutoMocker();
+
+			var studiedTexts = Enumerable.Range(1, 100)
+				.Select(n => CreateStudiedText($"{n}", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11).AddDays(n)))
+				.ToList();
+
 			var target = mocker.CreateInstance<TextsForPracticeSelector>();
 
 			// Act
 
-			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 21), studiedTexts);
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2024, 07, 11), studiedTexts, 100);
 
 			// Assert
 
-			var expectedTextsForCheck = new[]
-			{
-				studiedTexts[0],
-				studiedTexts[2],
-				studiedTexts[4],
-				studiedTexts[6],
-				studiedTexts[8],
-			};
+			var resultIds = textsForPractice.Select(x => Int32.Parse(x.TextInStudiedLanguage.Id.Value, NumberStyles.None, CultureInfo.InvariantCulture));
 
-			result.Should().BeEquivalentTo(expectedTextsForCheck, x => x.WithoutStrictOrdering());
+			resultIds.Should().NotBeInAscendingOrder();
 		}
 
 		[TestMethod]
-		public void GetTextsForPractice_ForTextsWithDifferentNextCheckDate_OrdersChecksByNextCheckDate()
+		public void GetTextsForPractice_IfOnlyTextsUnpracticedBeforeAreReturned_ReturnsTextsInRandomOrder()
 		{
 			// Arrange
 
-			var studiedTexts = new[]
-			{
-				new[]
-				{
-					// Next check: +1 day - 2023.07.11
-					GetFailedCheck(new DateTime(2023, 07, 10)),
-				}.ToStudiedText("Text with next check on 2023.07.11 - #1"),
-
-				// Next check: MinDate
-				Array.Empty<CheckResult>().ToStudiedText("Text with no checks - #1"),
-
-				new[]
-				{
-					// Next check: +1 day - 2023.07.10
-					GetFailedCheck(new DateTime(2023, 07, 09)),
-				}.ToStudiedText("Text with next check on 2023.07.10 - #1"),
-
-				new[]
-				{
-					// Next check: +1 day - 2023.07.11
-					GetFailedCheck(new DateTime(2023, 07, 10)),
-				}.ToStudiedText("Text with next check on 2023.07.11 - #2"),
-
-				// Next check: MinDate
-				Array.Empty<CheckResult>().ToStudiedText("Text with no checks - #2"),
-
-				new[]
-				{
-					// Next check: +1 day - 2023.07.10
-					GetFailedCheck(new DateTime(2023, 07, 09)),
-				}.ToStudiedText("Text with next check on 2023.07.10 - #2"),
-			};
-
 			var mocker = new AutoMocker();
+
+			var studiedTexts = Enumerable.Range(1, 100)
+				.Select(n => CreateStudiedTextWithNoChecks($"{n}", mocker, createDate: new DateTime(2023, 07, 11).AddDays(n)))
+				.ToList();
+
 			var target = mocker.CreateInstance<TextsForPracticeSelector>();
 
 			// Act
 
-			var result = target.GetTextsForPractice(new DateOnly(2023, 07, 11), studiedTexts);
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2024, 07, 11), studiedTexts, 100);
 
 			// Assert
 
-			result.Count.Should().Be(6);
+			var resultIds = textsForPractice.Select(x => Int32.Parse(x.TextInStudiedLanguage.Id.Value, NumberStyles.None, CultureInfo.InvariantCulture));
 
-			result.Skip(0).Take(2).Should().BeEquivalentTo(new[] { studiedTexts[1], studiedTexts[4] }, x => x.WithoutStrictOrdering());
-			result.Skip(2).Take(2).Should().BeEquivalentTo(new[] { studiedTexts[2], studiedTexts[5] }, x => x.WithoutStrictOrdering());
-			result.Skip(4).Take(2).Should().BeEquivalentTo(new[] { studiedTexts[0], studiedTexts[3] }, x => x.WithoutStrictOrdering());
+			resultIds.Should().NotBeInAscendingOrder();
 		}
 
-		private static CheckResult GetFailedCheck(DateTimeOffset dateTime)
+		[TestMethod]
+		public void GetTextsForPractice_IfPracticedAndUnpracticedTextsAreReturned_ReturnsTextsInRandomOrder()
 		{
-			return new CheckResult
+			// Arrange
+
+			var mocker = new AutoMocker();
+
+			var practicedTexts = Enumerable.Range(1, 100)
+				.Select(n => CreateStudiedText($"{n}", mocker, lastCheckDate: new DateTime(2023, 07, 10), nextCheckDate: new DateOnly(2023, 07, 11).AddDays(n)))
+				.ToList();
+
+			var unpracticedTexts = Enumerable.Range(101, 100)
+				.Select(n => CreateStudiedTextWithNoChecks($"{n}", mocker, createDate: new DateTime(2023, 07, 11).AddDays(n)))
+				.ToList();
+
+			var target = mocker.CreateInstance<TextsForPracticeSelector>();
+
+			// Act
+
+			var textsForPractice = target.GetTextsForPractice(new DateOnly(2024, 07, 11), practicedTexts.Concat(unpracticedTexts), 200);
+
+			// Assert
+
+			var resultIds = textsForPractice.Select(x => Int32.Parse(x.TextInStudiedLanguage.Id.Value, NumberStyles.None, CultureInfo.InvariantCulture)).ToList();
+
+			resultIds.Should().NotBeInAscendingOrder();
+			resultIds.Take(100).Should().NotBeInAscendingOrder();
+			resultIds.Skip(100).Take(100).Should().NotBeInAscendingOrder();
+		}
+
+		private static StudiedText CreateStudiedText(string id, AutoMocker mocker, DateTimeOffset lastCheckDate, DateOnly nextCheckDate)
+		{
+			var studiedText = new StudiedText(new[] { GetSuccessfulCheck(lastCheckDate) })
 			{
-				DateTime = dateTime,
-				CheckResultType = CheckResultType.Misspelled,
+				TextInStudiedLanguage = new LanguageText
+				{
+					Id = new ItemId(id),
+				},
 			};
+
+			mocker.GetMock<INextCheckDateProvider>()
+				.Setup(x => x.GetNextCheckDate(studiedText)).Returns(nextCheckDate);
+
+			return studiedText;
 		}
 
 		private static CheckResult GetSuccessfulCheck(DateTimeOffset dateTime)
@@ -406,6 +347,22 @@ namespace VocabularyCoach.Services.UnitTests.Internal
 				CheckResultType = CheckResultType.Ok,
 			};
 		}
-*/
+
+		private static StudiedText CreateStudiedTextWithNoChecks(string id, AutoMocker mocker, DateTimeOffset createDate)
+		{
+			var studiedText = new StudiedText(Enumerable.Empty<CheckResult>())
+			{
+				TextInStudiedLanguage = new LanguageText
+				{
+					Id = new ItemId(id),
+					CreationTimestamp = createDate,
+				},
+			};
+
+			mocker.GetMock<INextCheckDateProvider>()
+				.Setup(x => x.GetNextCheckDate(studiedText)).Returns(DateOnly.MinValue);
+
+			return studiedText;
+		}
 	}
 }