+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.8.0 - 2024.04.22
+- Added daily limit for practiced texts.
+- Enhanced selection of practiced texts for new users.
+- Added a new check interval of 21 days.
+
 ## 2.7.6 - 2024.03.04
 - Fixed application error if there are texts with no checks yet.